Barebones PDF generation
The client has a fake payment checkout page that sends hardcoded data to the server in return for a PDF. The PDF does not currently use the data, but a sample PDF is rendered on the page upon success.

Server is in a good starting point. I have built a route that takes json data and puts it into a PDF. This PDF is then encoded and sent back to the client. There is a flaw in my code somewhere, maybe caused by fetch? After 10 requests, the server spits out a message saying

"(node:14552) MaxListenersExceededWarning: Possible EventEmitter memory leak detected. 11 uncaughtException listeners added to [process]."

I've never seen this before and can't immediately think of what's causing it. The main difference is that I usually use axios but I'm using fetch here..